move cazy to dbs/cazy
